requirements.txt
python-dotenv>=0.19.1
requests>=2.24.0
bs4>=4.9.1
beautifulsoup4>=4.8.2